Updating build scripts

+++ prodbuild.sh
@@ -1,26 +1,37 @@
 #!/bin/bash
 
-\rm -f nginx-prod/build/*.tgz
-mkdir -p nginx-prod/build/
+CMD="$1"
 
-# build the frontend assets (this takes quite a while due to minification)
-docker-compose -f docker-compose-buildjs.yml run frontend
-docker-compose -f docker-compose-buildjs.yml stop
+if [ x"$CMD" = x ]; then
+    echo "provide a command!"
+    exit 1
+fi
 
-# build the django assets
-docker-compose -f docker-compose-buildpy.yml build
-docker-compose -f docker-compose-buildpy.yml run django
-docker-compose -f docker-compose-buildpy.yml stop
+if [ "$CMD" = "frontend" ] || [ "$CMD" = "all" ]; then
+  \rm -f nginx-prod/build/*.tgz
+  mkdir -p nginx-prod/build/
 
-# copy assets locally for build local production nginx image (local testing only)
-cp build/frontend.tgz build/django.tgz nginx-prod/build # this is horrible, fixme
+  # build the frontend assets (this takes quite a while due to minification)
+  docker-compose -f docker-compose-buildjs.yml run frontend
+  docker-compose -f docker-compose-buildjs.yml stop
 
-# For local testing with docker-compose-prod.yml only
-# echo building prod nginx container
-# (cd nginx-prod && docker build -t scremsong/nginx-prod:latest .)
-# (cd nginx-prod && docker build --no-cache -t scremsong/nginx-prod:latest . && cd ..)
+  # build the django assets
+  docker-compose -f docker-compose-buildpy.yml build
+  docker-compose -f docker-compose-buildpy.yml run django
+  docker-compose -f docker-compose-buildpy.yml stop
 
-echo building prod django container
-(cd django && docker build -t scremsong/django:latest .)
-# (cd django && docker build --no-cache -t scremsong/django:latest . && cd ..)
-# rm django/scremsong/ealfront/templates/index.html
+  # copy assets locally for build local production nginx image (local testing only)
+  cp build/frontend.tgz build/django.tgz nginx-prod/build # this is horrible, fixme
+
+  # For local testing with docker-compose-prod.yml only
+  # echo building prod nginx container
+  # (cd nginx-prod && docker build -t scremsong/nginx-prod:latest .)
+  # (cd nginx-prod && docker build --no-cache -t scremsong/nginx-prod:latest . && cd ..)
+fi
+
+if [ "$CMD" = "django" ] || [ "$CMD" = "all" ]; then
+  echo building prod django container
+  (cd django && docker build -t scremsong/django:latest .)
+  # (cd django && docker build --no-cache -t scremsong/django:latest . && cd ..)
+  # rm django/scremsong/ealfront/templates/index.html
+fi
